+++ src/Search.module.css
@@ -1,0 +1,17 @@
+.iconSearch {
+  color: var(--color-fg3);
+  position: absolute;
+  top: 8px;
+  left: 10px;
+}
+
+.iconClear {
+  color: var(--color-fg1);
+  position: absolute;
+  top: 6px;
+  right: 6px;
+}
+
+.hidden {
+  display: none;
+}
